+++ Sources/Swift/Span.swift
@@ -13,15 +13,5 @@ extension Span {
         }
         wrapper.buffer.add(name: name, value: result)
     }
-
-    @_spi(Private) public func removeFeatureFlag(name: String) {
-        guard let span = self as? SentrySpanInternal else {
-            return
-        }
-        guard let wrapper = span.featureFlagBuffer as? SentryFeatureFlagBufferWrapper else {
-            return
-        }
-        wrapper.buffer.remove(name: name)
-    }
 }
 // swiftlint:enable missing_docs